[심민성] 차량 관련 수정

+++ group_ware/src/main/java/com/ware/spring/vehicle/domain/VehicleSize.java
@@ -16,7 +16,7 @@ public class VehicleSize {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "size_no")
-    private int sizeNo;
+    private Long sizeNo;
 
     @Column(name = "vehicle_size", nullable = false)
     private String vehicleSize;
